requirements.txt
Flask==2.2.2
python_docx_replace==0.4.1